+++ CHANGELOG.md
@@ -3,4 +3,4 @@ here's only the changelog for the version in development
 
 # v1.2
 
-
+* removed the limit for string length and list and object size